+++ packages/backend-function/src/factory.ts
@@ -18,6 +18,8 @@ import {
   ConstructFactoryGetInstanceProps,
   FunctionResources,
   GenerateContainerEntryProps,
+  LogLevel,
+  LogRetention,
   ResourceAccessAcceptorFactory,
   ResourceNameValidator,
   ResourceProvider,
@@ -46,6 +48,7 @@ import { FunctionEnvironmentTranslator } from './function_env_translator.js';
 import { FunctionEnvironmentTypeGenerator } from './function_env_type_generator.js';
 import { FunctionLayerArnParser } from './layer_parser.js';
 import { convertFunctionSchedulesToRuleSchedules } from './schedule_parser.js';
+import { convertLoggingOptionsToCDK } from './logging_options_parser.js';
 import {
   ProvidedFunctionFactory,
   ProvidedFunctionProps,
@@ -68,6 +71,9 @@ export type TimeInterval =
   | `every month`
   | `every year`;
 export type FunctionSchedule = TimeInterval | CronSchedule;
+
+export type FunctionLogLevel = LogLevel;
+export type FunctionLogRetention = LogRetention;
 
 export function defineFunction(
   props?: FunctionProps
@@ -185,6 +191,8 @@ export type FunctionProps = {
    * resourceGroupName: 'auth' // to group an auth trigger with an auth resource
    */
   resourceGroupName?: AmplifyResourceGroupName;
+
+  logging?: FunctionLoggingOptions;
 };
 
 export type FunctionBundlingOptions = {
@@ -194,6 +202,18 @@ export type FunctionBundlingOptions = {
    * Defaults to true.
    */
   minify?: boolean;
+};
+
+export type FunctionLoggingOptions = (
+  | {
+      format: 'json';
+      level?: FunctionLogLevel;
+    }
+  | {
+      format?: 'text';
+    }
+) & {
+  retention?: FunctionLogRetention;
 };
 
 /**
@@ -238,12 +258,13 @@ class FunctionFactory implements ConstructFactory<AmplifyFunction> {
       entry: this.resolveEntry(),
       timeoutSeconds: this.resolveTimeout(),
       memoryMB: this.resolveMemory(),
-      environment: this.props.environment ?? {},
+      environment: this.resolveEnvironment(),
       runtime: this.resolveRuntime(),
       schedule: this.resolveSchedule(),
       bundling: this.resolveBundling(),
       layers,
       resourceGroupName: this.props.resourceGroupName ?? 'function',
+      logging: this.props.logging ?? {},
     };
   };
 
@@ -299,9 +320,10 @@ class FunctionFactory implements ConstructFactory<AmplifyFunction> {
         timeoutMax
       )
     ) {
-      throw new Error(
-        `timeoutSeconds must be a whole number between ${timeoutMin} and ${timeoutMax} inclusive`
-      );
+      throw new AmplifyUserError('InvalidTimeoutError', {
+        message: `Invalid function timeout of ${this.props.timeoutSeconds}`,
+        resolution: `timeoutSeconds must be a whole number between ${timeoutMin} and ${timeoutMax} inclusive`,
+      });
     }
     return this.props.timeoutSeconds;
   };
@@ -316,11 +338,39 @@ class FunctionFactory implements ConstructFactory<AmplifyFunction> {
     if (
       !isWholeNumberBetweenInclusive(this.props.memoryMB, memoryMin, memoryMax)
     ) {
-      throw new Error(
-        `memoryMB must be a whole number between ${memoryMin} and ${memoryMax} inclusive`
-      );
+      throw new AmplifyUserError('InvalidMemoryMBError', {
+        message: `Invalid function memoryMB of ${this.props.memoryMB}`,
+        resolution: `memoryMB must be a whole number between ${memoryMin} and ${memoryMax} inclusive`,
+      });
     }
     return this.props.memoryMB;
+  };
+
+  private resolveEnvironment = () => {
+    if (this.props.environment === undefined) {
+      return {};
+    }
+
+    const invalidKeys: string[] = [];
+
+    Object.keys(this.props.environment).forEach((key) => {
+      // validate using key pattern from https://docs.aws.amazon.com/lambda/latest/api/API_Environment.html
+      if (!key.match(/^[a-zA-Z]([a-zA-Z0-9_])+$/)) {
+        invalidKeys.push(key);
+      }
+    });
+
+    if (invalidKeys.length > 0) {
+      throw new AmplifyUserError('InvalidEnvironmentKeyError', {
+        message: `Invalid function environment key(s): ${invalidKeys.join(
+          ', '
+        )}`,
+        resolution:
+          'Environment keys must match [a-zA-Z]([a-zA-Z0-9_])+ and be at least 2 characters',
+      });
+    }
+
+    return this.props.environment;
   };
 
   private resolveRuntime = () => {
@@ -332,11 +382,12 @@ class FunctionFactory implements ConstructFactory<AmplifyFunction> {
     }
 
     if (!(this.props.runtime in nodeVersionMap)) {
-      throw new Error(
-        `runtime must be one of the following: ${Object.keys(
+      throw new AmplifyUserError('InvalidRuntimeError', {
+        message: `Invalid function runtime of ${this.props.runtime}`,
+        resolution: `runtime must be one of the following: ${Object.keys(
           nodeVersionMap
-        ).join(', ')}`
-      );
+        ).join(', ')}`,
+      });
     }
 
     return this.props.runtime;
@@ -464,6 +515,7 @@ class AmplifyFunction
     functionEnvironmentTypeGenerator.generateProcessEnvShim();
 
     let functionLambda: NodejsFunction;
+    const cdkLoggingOptions = convertLoggingOptionsToCDK(props.logging);
     try {
       functionLambda = new NodejsFunction(scope, `${id}-lambda`, {
         entry: props.entry,
@@ -477,6 +529,9 @@ class AmplifyFunction
           inject: shims,
           externalModules: Object.keys(props.layers),
         },
+        logRetention: cdkLoggingOptions.retention,
+        applicationLogLevelV2: cdkLoggingOptions.level,
+        loggingFormat: cdkLoggingOptions.format,
       });
     } catch (error) {
       // If the error is from ES Bundler which is executed as a child process by CDK,