+++ packages/cli/src/commands/pipeline-deploy/pipeline_deploy_command.test.ts
@@ -238,4 +238,32 @@ void describe('deploy command', () => {
       ClientConfigFormat.DART,
     ]);
   });
+
+  void it('throws when --branch argument has no input', async () => {
+    await assert.rejects(
+      async () =>
+        await getCommandRunner(true).runCommand(
+          'pipeline-deploy --app-id abc --branch'
+        ),
+      (error: TestCommandError) => {
+        assert.strictEqual(error.error.name, 'InvalidCommandInputError');
+        assert.strictEqual(error.error.message, 'Invalid --branch or --app-id');
+        return true;
+      }
+    );
+  });
+
+  void it('throws when --app-id argument has no input', async () => {
+    await assert.rejects(
+      async () =>
+        await getCommandRunner(true).runCommand(
+          'pipeline-deploy --app-id --branch testBranch'
+        ),
+      (error: TestCommandError) => {
+        assert.strictEqual(error.error.name, 'InvalidCommandInputError');
+        assert.strictEqual(error.error.message, 'Invalid --branch or --app-id');
+        return true;
+      }
+    );
+  });
 });