+++ packages/integration-tests/src/process-controller/execa_process_killer.ts
@@ -1,10 +1,12 @@
-import { ExecaChildProcess, execa } from 'execa';
+import { ExecaMethod, execa } from 'execa';
 
 /**
  * Kills the given process (equivalent of sending CTRL-C)
  * @param processInstance an instance of execa child process
  */
-export const killExecaProcess = async (processInstance: ExecaChildProcess) => {
+export const killExecaProcess = async (
+  processInstance: ReturnType<ExecaMethod>
+) => {
   if (process.platform.startsWith('win')) {
     if (typeof processInstance.pid !== 'number') {
       throw new Error('Cannot kill the process that does not have pid');