+++ packages/integration-tests/src/test-e2e/sandbox/data_access_from_function_project.sandbox.test.ts
@@ -1,0 +1,4 @@
+import { defineSandboxTest } from './sandbox.test.template.js';
+import { DataAccessFromFunctionTestProjectCreator } from '../../test-project-setup/data_access_from_function_project.js';
+
+defineSandboxTest(new DataAccessFromFunctionTestProjectCreator());